django-projects/cash3/requirements.txt
asgiref==3.4.1
black==21.8b0
blank==0.1.0
click==8.0.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-graphos==0.3.41
django-heroku==0.3.1
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.3.0
psycopg2==2.9.1
pytz==2021.1
regex==2021.8.28
six==1.16.0
sqlparse==0.4.1
tomli==1.2.1
typing-extensions==3.10.0.2
whitenoise==5.3.0
